Renames HiveUtils to HiveCompat
