+++ sql/hive/v0.13.1/src/main/scala/org/apache/spark/sql/hive/Shim13.scala
@@ -156,7 +156,7 @@ object HiveShim {
     } else {
       // TODO precise, scale?
       new hiveIo.HiveDecimalWritable(
-        HiveUtils.newHiveDecimal(value.asInstanceOf[Decimal].toJavaBigDecimal))
+        HiveCompat.newHiveDecimal(value.asInstanceOf[Decimal].toJavaBigDecimal))
     }
 
   def createDriverResultsArray = new JArrayList[Object]
